+++ packages/showcases/core/csf/components/selection-controls/quantity.csf.js
@@ -7,7 +7,7 @@ export const parameters = {
       'focusout',
       'add',
       'subtract',
-      'blur',
+      'change',
     ],
   },
   design: {